Add files via upload
Description

The heatmap visualizes the presence of null (missing) values in the dataset used for the SONAR Rock and Mine Prediction project. The uniform dark purple color across the plot indicates that there are no missing values in the dataset. This suggests that:

The dataset is complete and does not require any imputation.

No preprocessing steps related to handling missing values are necessary.

The machine learning models can be trained without concerns about data gaps.